split SubscriptionFactory into .cc and .h
